Move resolvers module to top-level
