+++ crates/apollo-compiler/src/execution/engine.rs
@@ -5,14 +5,14 @@ use crate::collections::IndexMap;
 use crate::executable::Field;
 use crate::executable::Selection;
 use crate::execution::input_coercion::coerce_argument_values;
-use crate::execution::resolver::ObjectValue;
-use crate::execution::resolver::ResolveError;
-use crate::execution::resolver::ResolvedValue;
 use crate::execution::result_coercion::complete_value;
 use crate::introspection::resolvers::MaybeLazy;
 use crate::introspection::resolvers::SchemaWithImplementersMap;
 use crate::parser::SourceMap;
 use crate::parser::SourceSpan;
+use crate::resolvers::ObjectValue;
+use crate::resolvers::ResolveError;
+use crate::resolvers::ResolvedValue;
 use crate::response::GraphQLError;
 use crate::response::JsonMap;
 use crate::response::JsonValue;
